refactor: 메인 뷰 모델 이식 및 기존 연결 교체
